# backend/app/oa/service/cmd/server/assets/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
    /admin/v1/captcha:
        get:
            tags:
                - AuthenticationService
            description: 生成验证码。转发至 cms /admin/v1/captcha。
            operationId: AuthenticationService_GenerateCaptcha
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/GenerateCaptchaResponse'
            security:
                - {}
    /admin/v1/captcha/verify:
        post:
            tags:
                - AuthenticationService
            description: 验证验证码。转发至 cms /admin/v1/captcha/verify。
            operationId: AuthenticationService_VerifyCaptcha
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/VerifyCaptchaRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/VerifyCaptchaResponse'
            security:
                - {}
    /admin/v1/login:
        post:
            tags:
                - AuthenticationService
            description: 登录。转发至 cms /admin/v1/login。
            operationId: AuthenticationService_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
            security:
                - {}
    /admin/v1/logout:
        post:
            tags:
                - AuthenticationService
            description: 登出。转发至 cms /admin/v1/logout。
            operationId: AuthenticationService_Logout
            requestBody:
                content:
                    application/json: {}
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
...
    /admin/v1/refresh-token:
        post:
            tags:
                - AuthenticationService
            description: 刷新认证令牌。转发至 cms /admin/v1/refresh-token。
            operationId: AuthenticationService_RefreshToken
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LoginResponse'
components:
    schemas:
...
        GenerateCaptchaResponse:
            type: object
            properties:
                captchaId:
                    type: string
                    description: 验证码ID，客户端应在后续请求中提供该ID以验证用户输入的验证码
                imageBase64:
                    type: string
                    description: 验证码图片的Base64编码字符串，客户端可以解码并显示给用户
            description: 生成验证码 - 回应。与 cms GenerateCaptchaResponse 同构。
...
        LoginRequest:
            required:
                - grant_type
            type: object
            properties:
                grant_type:
                    enum:
                        - password
                        - client_credentials
                        - authorization_code
                        - refresh_token
                        - implicit
                    type: string
                    default: password
                    description: 授权类型，此处的值固定为"password"，必选项。
                    format: enum
                client_id:
                    type: string
                    description: 客户端ID
                client_secret:
                    type: string
                    description: 客户端密钥
                scope:
                    type: string
                    description: 以空格分隔的用户授予范围列表。如果未提供，scope则授权任何范围，默认为空列表。
                redirect_uri:
                    type: string
                    description: 跳转链接
                user_id:
                    type: integer
                    description: 用户ID
                    format: uint32
                username:
                    type: string
                    description: 用户名
                email:
                    type: string
                    description: 电子邮件地址
                mobile:
                    type: string
                    description: 手机号
                password:
                    type: string
                    description: 用户的密码
                refresh_token:
                    type: string
                    description: 更新令牌，用来获取下一次的访问令牌，可选项。如果访问令牌将过期，则返回刷新令牌很有用，应用程序可以使用该刷新令牌获取另一个访问令牌。但是，通过隐式授予颁发的令牌不能颁发刷新令牌。
                code:
                    type: string
                    description: 授权请求中收到的一次性验证/认证码。(当使用授权码模式时)
                client_type:
                    enum:
                        - admin
                        - app
                    type: string
                    description: 客户端类型
                    format: enum
                device_id:
                    type: string
                    description: 设备唯一标识（可选），用于设备绑定、推送、风控等
                jti:
                    type: string
                    description: 建议客户端生成并提供 jti（JWT ID）作为唯一标识，服务端可据此防止重放攻击
                tenant_code:
                    type: string
                    description: 租户编号，留空表示平台登录（平台超级管理员）；非空时按该编号解析对应租户
            description: |-
                用户登录 - 请求。与 cms LoginRequest 同构（字段编号、json_name、OpenAPI
                 描述逐字段对齐），仅剥离 (redact.value) 注解。
        LoginResponse:
            type: object
            properties:
                token_type:
                    enum:
                        - bearer
                        - mac
                    type: string
                    default: Bearer
                    description: 令牌的类型，该值大小写不敏感，必选项，可以是bearer类型或mac类型，通常只是字符串“Bearer”。
                    format: enum
                access_token:
                    type: string
                    description: 访问令牌，必选项。授权服务器颁发的访问令牌字符串。
                expires_in:
                    type: string
                    description: 访问令牌过期时间（秒）
                refresh_token:
                    type: string
                    description: 更新令牌，用来获取下一次的访问令牌，可选项。如果访问令牌将过期，则返回刷新令牌很有用，应用程序可以使用该刷新令牌获取另一个访问令牌。但是，通过隐性授予颁发的令牌不能颁发刷新令牌。
                scope:
                    type: string
                    description: 以空格分隔的用户授予范围列表。如果未提供，scope则授权任何范围，默认为空列表。
                refresh_expires_in:
                    type: string
                    description: 刷新令牌过期时间（秒）
                id_token:
                    type: string
                    description: ID 令牌，OpenID Connect 扩展中定义的 JWT 格式令牌
            description: 用户登录 - 回应。与 cms LoginResponse 同构。
...
        VerifyCaptchaRequest:
            type: object
            properties:
                captchaId:
                    type: string
                    description: 验证码ID，来自 GenerateCaptcha 响应
                userInput:
                    type: string
                    description: 用户输入的验证码文本
            description: 验证验证码 - 请求。与 cms VerifyCaptchaRequest 同构。
        VerifyCaptchaResponse:
            type: object
            properties:
                valid:
                    type: boolean
                    description: 验证码验证结果，true表示验证成功，false表示验证失败
            description: 验证验证码 - 回应。与 cms VerifyCaptchaResponse 同构。
...